Move ibc and ibc_test to internal/ibc. They are helper functions that can be used elsewhere.
