+++ x/ibcratelimit/module/ibc_middleware.go
@@ -13,6 +13,7 @@ import (
 	porttypes "github.com/cosmos/ibc-go/v6/modules/core/05-port/types"
 	"github.com/cosmos/ibc-go/v6/modules/core/exported"
 
+	"github.com/provenance-io/provenance/internal/ibc"
 	"github.com/provenance-io/provenance/x/ibcratelimit"
 	"github.com/provenance-io/provenance/x/ibcratelimit/keeper"
 )
@@ -129,7 +130,7 @@ func (im *IBCMiddleware) OnRecvPacket(
 	relayer sdk.AccAddress,
 ) exported.Acknowledgement {
 	if err := ibcratelimit.ValidateReceiverAddress(packet); err != nil {
-		return ibcratelimit.NewEmitErrorAcknowledgement(ctx, ibcratelimit.ErrBadMessage, err.Error())
+		return ibc.NewEmitErrorAcknowledgement(ctx, ibcratelimit.ErrBadMessage, err.Error())
 	}
 
 	if !im.keeper.ContractConfigured(ctx) {
@@ -139,7 +140,7 @@ func (im *IBCMiddleware) OnRecvPacket(
 
 	err := im.keeper.CheckAndUpdateRateLimits(ctx, "recv_packet", packet)
 	if err != nil {
-		return ibcratelimit.NewEmitErrorAcknowledgement(ctx, err)
+		return ibc.NewEmitErrorAcknowledgement(ctx, err)
 	}
 
 	// if this returns an Acknowledgement that isn't successful, all state changes are discarded
@@ -158,7 +159,7 @@ func (im *IBCMiddleware) OnAcknowledgementPacket(
 		return errorsmod.Wrapf(sdkerrors.ErrUnknownRequest, "cannot unmarshal ICS-20 transfer packet acknowledgement: %v", err)
 	}
 
-	if ibcratelimit.IsAckError(acknowledgement) {
+	if ibc.IsAckError(acknowledgement) {
 		err := im.keeper.RevertSentPacket(ctx, packet) // If there is an error here we should still handle the ack
 		if err != nil {
 			eventErr := ctx.EventManager().EmitTypedEvent(&ibcratelimit.EventAckRevertFailure{
